% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslices.R
\name{dtm2tsl}
\alias{dtm2tsl}
\title{Convert date-time objects to time-slice}
\usage{
dtm2tsl(dtm, format = "d365_h24", d366.as.na = grepl("d365", format))
}
\arguments{
\item{dtm}{vector of timepoints in Date format}

\item{format}{character, format of the slices}

\item{d366.as.na}{logical, if}
}
\value{
Character vector with time-slices names
}
\description{
Convert date-time objects to time-slice
}
\examples{
dtm2tsl(lubridate::now())
dtm2tsl(lubridate::ymd("2020-12-31"))
dtm2tsl(lubridate::ymd("2020-12-31"), d366.as.na = F)
dtm2tsl(lubridate::now(tzone = "UTC"), format = "d365")
dtm2tsl(lubridate::ymd("2020-12-31"), format = "d365")
dtm2tsl(lubridate::ymd("2020-12-31"), format = "d365", d366.as.na = F)
dtm2tsl(lubridate::ymd("2020-12-31"), format = "d366")
}
