% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslices-package.R
\docType{package}
\name{timeslices-package}
\alias{timeslices}
\alias{timeslices-package}
\title{timeslices: Time-slices for energy models}
\description{
Most common set of time-slices with functions to convert from date-time format to strings and back.
}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@gmail.com} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

}
\keyword{internal}
