% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslices.R
\name{tsl2dtm}
\alias{tsl2dtm}
\alias{tsl2year}
\alias{tsl2yday}
\alias{tsl2hour}
\alias{tsl2month}
\title{Convert time-slices to date-time, extract year, month, day of the year, or hour}
\usage{
tsl2dtm(
  tsl,
  type = tsl_guess_format(tsl),
  tmz = "UTC",
  year = NULL,
  mday = NULL
)

tsl2year(tsl, return.null = T)

tsl2yday(tsl, return.null = T)

tsl2hour(tsl, return.null = T)

tsl2month(tsl, return.null = T)
}
\arguments{
\item{tsl}{character vector with time-slices}

\item{type}{character, type of the slices}

\item{tmz}{time-zone}

\item{year}{year, used when time-slices don't store year}

\item{mday}{day of month, for time slices without the information}

\item{return.null}{logical, valid for the cased then all values are NA, then NULL will be returned if return.null = TRUE,}
}
\value{
Vector in Date-Time format








}
\description{
Convert time-slices to date-time, extract year, month, day of the year, or hour
}
\section{Functions}{
\itemize{
\item \code{tsl2year}: Extract year from time-slices

\item \code{tsl2yday}: Extract the day of the year from time-slices

\item \code{tsl2hour}: Extract hour from time-slices

\item \code{tsl2month}: Extract month from time-slices
}}

\examples{
tsl <- c("y2007_d365_h15", NA, "d151_h22", "d001", "m10_h12")
tsl2dtm(tsl[1])
tsl2dtm(tsl[1:2])
tsl2dtm(tsl[2])
tsl2dtm(tsl[3])
tsl2dtm(tsl[4])
tsl2dtm(tsl[3], year = 2010)
tsl2dtm(tsl[4], year = 1900)
tsl2dtm(tsl[3:4], year = 1900)
NULL
NULL
NULL
NULL
}
