% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslices.R
\docType{data}
\name{tsl_formats}
\alias{tsl_formats}
\title{Common formats of time-slices.}
\format{
A character vector with formats:
\describe{
\item{d365}{daily time-slices, 365 a year (leap year's 366th day is disregarded)}
\item{d365_h24}{time slices with year-day numbers and hours, 8760 in total}
...
}
}
\usage{
tsl_formats
}
\description{
Common formats of time-slices.
}
\keyword{datasets}
